build!: remove unnecessary module dependencies
This change removes all existing dependencies from package.json, since
the svelte-masonry module itself (Masonry.svelte) has no external
dependencies. This change allows this module to be used without
modification as a dependency in node 14, which was previously not
possible due to the node-sass dependency. The engine (node 12.x) was
also removed so that npm won't complain under node 14.

BREAKING CHANGE: This completely breaks the build of the demo app.
Personally, my recommendation would be to break out that app into a
separate package, which will keep the svelte-masonry package size small
and avoid the problem of these unnecessary dependencies.

This change shouldn't affect users of svelte-masonry, since the only
useful code is in Masonry.svelte. A future commit will remove the
remaining demo app code.